third_party/WebKit/LayoutTests/fast/scrolling/scrolling-apis-subpixel-expected.txt

Verifies that scrolling APIs support fractional offsets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


---- Testing page zoom = 2 ----
Scrolling DIV with scrollTop/scrollLeft
PASS scroller.scrollTop is 4
PASS scroller.scrollLeft is 4
Scrolling the document with window.scroll
PASS window.scrollY is 5
PASS window.scrollX is 5
Scrolling the document with window.scrollTo
PASS window.pageYOffset is 5
PASS window.pageXOffset is 5
Scrolling the document with window.scrollBy
PASS window.scrollY is 5
PASS window.scrollX is 5

Scrolling DIV with scrollTop/scrollLeft
PASS scroller.scrollTop is 4.5
PASS scroller.scrollLeft is 4.5
Scrolling the document with window.scroll
PASS window.scrollY is 5.5
PASS window.scrollX is 5.5
Scrolling the document with window.scrollTo
PASS window.pageYOffset is 5.5
PASS window.pageXOffset is 5.5
Scrolling the document with window.scrollBy
PASS window.scrollY is 5.5
PASS window.scrollX is 5.5

PASS successfullyParsed is true

TEST COMPLETE
